wip - show bookmarks in modal
