+++ src/frontend/components/AppHeader/index.tsx
@@ -71,7 +71,7 @@ export default function AppHeader() {
     },
     {
       label: 'Bookmarks',
-      onClick: () => navigate('/bookmarks'),
+      onClick: () => selectCommand({ event: 'clientEvent/bookmark/show' }), // () => navigate('/bookmarks')
       startIcon: <StarIcon />,
     },
     {
